optimization: template export_template api unified
