+++ gcloud/template_base/apis/django/api.py
@@ -11,11 +11,18 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express o
 specific language governing permissions and limitations under the License.
 """
 
-# unified_api_utils.py：项目流程和公共流程api接口的统一实现，减少重复代码
+import ujson as json
+import hashlib
+import base64
+
 from django.http import HttpRequest
-from django.http import JsonResponse
+from django.http import JsonResponse, HttpResponse
+
 
 from gcloud import err_code
+from gcloud.conf import settings
+from gcloud.utils.dates import time_now_str
+from gcloud.exceptions import FlowExportError
 from gcloud.template_base.utils import read_template_data_file
 
 
@@ -80,3 +87,30 @@ def base_check_before_import(request: HttpRequest, template_model_cls: object, i
     check_info = template_model_cls.objects.import_operation_check(r["data"]["template_data"], *import_args)
 
     return JsonResponse({"result": True, "data": check_info, "code": err_code.SUCCESS.code, "message": ""})
+
+
+def base_export_templates(request: HttpRequest, template_model_cls: object, file_prefix: str, export_args: list):
+    data = json.loads(request.body)
+    template_id_list = data["template_id_list"]
+
+    # wash
+    try:
+        templates_data = json.loads(
+            json.dumps(template_model_cls.objects.export_templates(template_id_list, *export_args), sort_keys=True)
+        )
+    except FlowExportError as e:
+        return JsonResponse({"result": False, "message": str(e), "code": err_code.UNKNOWN_ERROR.code, "data": None})
+
+    data_string = (json.dumps(templates_data, sort_keys=True) + settings.TEMPLATE_DATA_SALT).encode("utf-8")
+    digest = hashlib.md5(data_string).hexdigest()
+
+    file_data = base64.b64encode(
+        json.dumps({"template_data": templates_data, "digest": digest}, sort_keys=True).encode("utf-8")
+    )
+    filename = "bk_sops_%s_%s.dat" % (file_prefix, time_now_str())
+    response = HttpResponse()
+    response["Content-Disposition"] = "attachment; filename=%s" % filename
+    response["mimetype"] = "application/octet-stream"
+    response["Content-Type"] = "application/octet-stream"
+    response.write(file_data)
+    return response
