+++ gcloud/template_base/apis/django/validators.py
@@ -11,7 +11,7 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express o
 specific language governing permissions and limitations under the License.
 """
 
-from gcloud.utils.validate import RequestValidator
+from gcloud.utils.validate import RequestValidator, ObjectJsonBodyValidator
 
 
 class BatchFormValidator(RequestValidator):
@@ -33,5 +33,21 @@ class FormValidator(RequestValidator):
     def validate(self, request, *args, **kwargs):
         if not request.GET.get("template_id"):
             return False, "template_id can not be empty"
+
+        return True, ""
+
+
+class ExportTemplateValidator(ObjectJsonBodyValidator):
+    def validate(self, request, *args, **kwargs):
+
+        super().validate(request, *args, **kwargs)
+
+        template_id_list = self.data.get("template_id_list")
+
+        if not isinstance(template_id_list, list):
+            return False, "invalid template_id_list"
+
+        if not template_id_list:
+            return False, "template_id_list can not be empty"
 
         return True, ""
